POCImplementations/AIChatbot/backend/python/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
anthropic>=0.18.0
openai==1.3.5
httpx==0.25.2
sqlalchemy==2.0.23
aiosqlite==0.19.0
pyyaml==6.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1

